+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- [#3707](https://github.com/ignite/cli/pull/3707) Add collections support.
 - [#4019](https://github.com/ignite/cli/pull/4019) Add `skip-proto` flag to `s chain` command
 - [#3977](https://github.com/ignite/cli/pull/3977) Add `chain lint` command to lint the chain's codebase using `golangci-lint`
 - [#3770](https://github.com/ignite/cli/pull/3770) Add `scaffold configs` and `scaffold params` commands